+++ src/main/java/com/greatmancode/legendarybotapi/item/ItemImpl.java
@@ -2,11 +2,11 @@ package com.greatmancode.legendarybotapi.item;
 
 public class ItemImpl implements Item {
 
-    private int id;
+    private long id;
     private String json;
 
-    public int getid() {return id;}
-    public void setid(int id) {this.id = id;}
+    public long getid() {return id;}
+    public void setid(long id) {this.id = id;}
 
     public String getJson() {return json;}
     public void setJson(String json) {this.json = json;}
