+++ src/main/java/com/greatmancode/legendarybotapi/utils/GenericIdJsonTable.java
@@ -2,8 +2,8 @@ package com.greatmancode.legendarybotapi.utils;
 
 public interface GenericIdJsonTable {
 
-    int getid();
-    void setid(int id);
+    long getid();
+    void setid(long id);
 
     String getJson();
     void setJson(String json);
